GCM Switch will later change to Firebase
Polling is when the device regularly calls up the server to see if
anything has changed. Pushing is when the server calls the device to say
something has changed. Polling has a lesser implementation cost, but can
end up wasting a lot of bandwidth and networking effort to check in with
a server that doesn't change very often.